// src/types.ts
export interface params_schema {
  top_islands: number;
  min_price: number;
  max_price: number;
}

export interface island_schema {
  background: string;
  category: string;
  clown: number;
  creationTime: string;
  description: string;
  discordOnly: number;
  fee: number;
  fruit: string;
  heart: number;
  hemisphere: string;
  islandScore: number;
  islandTime: string;
  islander: string;
  live: number;
  maxQueue: number;
  messageID: string;
  name: string;
  patreon: number;
  patreonOnly: number;
  poop: number;
  queued: string;
  rating: number;
  ratingCount: number;
  thumbsdown: number;
  thumbsupt: number;
  turnipCode: string;
  turnipPrice: number;
  watchlist: number;
}